Okay so this almost works.
Added another function to run motor continuously until scan is finished
and splice out odd indexes of barcode because they are unnecessary. Also
changes readings of the number 6 to zeros because the internet said this
was the color sensor value for white and idk why it just now started
spitting out 6es, but its weird so I switched them to zeros.